==> 1_Data_Cleaning (1).ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv('LoanApprovalPrediction.csv')

df1 = df.copy()    # copy created, incase needed

df.head(5)

df.info()

# null values in entire dataset
print(df.isnull().sum().sum())

df.isnull().sum()

(df.isnull().sum()/df.shape[0])*100

<h3>Dealing with Null Values</h3>

# Dependents → 2.0% (very low, safe to impute with mode/most common value)
df['Dependents'] = df['Dependents'].fillna(df['Dependents'].mode()[0])

# LoanAmount → 3.5% (numerical, can be imputed with median because income/loan data often has outliers)
df['LoanAmount'] = df['LoanAmount'].fillna(df['LoanAmount'].median())

# Loan_Amount_Term → 2.3% (categorical/numeric, usually 360 months, so you can impute with mode)
df['Loan_Amount_Term'] = df['Loan_Amount_Term'].fillna(df['Loan_Amount_Term'].mode()[0])

# Credit_History → 8.2% (highest missing, critical feature, you should not drop it. Impute with mode = 1, since majority usually have credit history)
df['Credit_History'] = df['Credit_History'].fillna(df['Credit_History'].mode()[0])

<p>All Null values are handled carefully without droping.</p>

<p>No duplicates present in dataset.</p>

df.to_csv('Cleaned_Loaned.csv',index=False)
