==> 7_SVM.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import MinMaxScaler
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
from sklearn.svm import SVC

df = pd.read_csv('Final_Scale_Encoded_Loan_data.csv')

X = df[['Gender_Encode','Married_Encode','Education_Encode','Self_Employed_Encode', 'Dependents', 'ApplicantIncome',
       'CoapplicantIncome', 'LoanAmount', 'Loan_Amount_Term', 'Credit_History','Property_Area_Encode']]

y = df['Loan_Status_Encode']

X_train, X_test, y_train, y_test= train_test_split(X,y,test_size=0.3,random_state=42)

scale_cols = ['ApplicantIncome', 'CoapplicantIncome', 'LoanAmount', 'Loan_Amount_Term']

scaler = MinMaxScaler()

X_train[scale_cols] = scaler.fit_transform(X_train[scale_cols])
X_test[scale_cols] = scaler.transform(X_test[scale_cols])

models = {
    'Linear_SVM': SVC(kernel='linear'),
    'Poly2_SVM' : SVC(kernel='poly',degree=2),
    'Pol3_SVM'  : SVC(kernel='poly',degree=3),
    'RBF_SVM'   : SVC(kernel='rbf')
}

for name,model in models.items():
    model.fit(X_train,y_train)
    train_score = round(model.score(X_train,y_train)*100,2)
    test_score = round(model.score(X_test,y_test)*100,2)
    print(name,train_score,test_score)

model1 = {
    'Linear_SVM': SVC(kernel='linear'),
    'Poly2_SVM' : SVC(kernel='poly',degree=2),
    'Pol3_SVM'  : SVC(kernel='poly',degree=3),
    'RBF_SVM'   : SVC(kernel='rbf')
}

for name,model in model1.items():
    model.fit(X_train,y_train)
    train_score = round(model.score(X_train,y_train)*100,2)
    y_pred = model.predict(X_test)
    test_score = round(accuracy_score(y_test,y_pred)*100,2)
    prec_test = round(precision_score(y_test,y_pred)*100,2)
    recall_test = round(recall_score(y_test,y_pred)*100,2)
    f1_test = round(f1_score(y_test,y_pred)*100,2)
    print(name,train_score,test_score,prec_test,recall_test,f1_test)

<h5>Poly3 SVM is your winner (best balance overall).</h5>

SVM_MODEL = SVC(kernel='poly',degree=3)
SVM_MODEL.fit(X_train,y_train)

print("Training data score of SVM model is",round(SVM_MODEL.score(X_train,y_train)*100,2))

y_test_pred = SVM_MODEL.predict(X_test)

print("Testing data score of SVM model is",round(accuracy_score(y_test,y_test_pred)*100,2))

<h1>Accuracy Score</h1>

<h4>Training data score of SVM model is 82.06</h4>

<h4>Testing data score of SVM model is 80.0</h4>

print("Accuracy :", round(accuracy_score(y_test, y_test_pred)*100, 2))
print("Precision:", round(precision_score(y_test, y_test_pred)*100, 2))
print("Recall   :", round(recall_score(y_test, y_test_pred)*100, 2))
print("F1-score :", round(f1_score(y_test, y_test_pred)*100, 2))
